+++ pynopegl/_pynopegl.pyx
@@ -115,7 +115,7 @@ cdef extern from "nopegl.h":
         const char *name
         int is_default
         ngl_cap *caps
-        int nb_caps
+        size_t nb_caps
 
     cdef struct ngl_scene:
         void *internal
